+++ src/__init__.py
@@ -6,3 +6,12 @@
 
 # import hysteresis.hysteresis
 # __all__ = ["hysteresis"]
+
+
+
+from .baseClass import Hysteresis, SimpleCycle, MonotonicCurve
+from .baseFuncs import concatenateHys, getReturnCycle, removeNegative
+from .resample import resample, resampleDx
+
+from .compare import compareCycle, compareHys
+from .protocol import exandHysTrace, createProtocol
